+++ src/SecucardConnect/SecucardConnect.php
@@ -33,7 +33,7 @@ final class SecucardConnect
     /**
      * SDK version
      */
-    const VERSION = '1.17.1';
+    const VERSION = '1.18.0';
 
     /**
      * @var OAuthProvider